Moved reward files
Moved files from RewardCollection to data folder and to reward-component folder. Renamed some files. Renamed giants data file to GiantsData so it's more clear what is used elsewhere in the code.
