+++ components/Reward/Reward.js
