+++ components/Reward/RewardStyles.js
@@ -1,9 +1,21 @@
 import { StyleSheet } from "react-native";
 import { widthPercentageToDP as wp, heightPercentageToDP as hp } from "react-native-responsive-screen";
-import Colors from "../../constants/colors";
 import { RFPercentage } from "react-native-responsive-fontsize";
+import Colors from "../../constants/colors";
 
 const styles = StyleSheet.create({
+  reward_big: {
+    width: 130,
+    height: 130,
+    marginTop: "60%",
+    marginLeft: "62%"
+  },
+  reward_small: {
+    width: 70,
+    height: 70,
+    marginTop: "77%",
+    marginLeft: "77%"
+  },
   container: {
     alignItems: "center",
     justifyContent: "center",
