--- components/Reward/styles.js
@@ -1,19 +1,0 @@
-import { StyleSheet } from "react-native";
-
-const styles = StyleSheet.create({
-
-  reward_big: {
-    width: 130,
-    height: 130,
-    marginTop: "60%",
-    marginLeft: "62%"
-  },
-  reward_small: {
-    width: 70,
-    height: 70,
-    marginTop: "77%",
-    marginLeft: "77%"
-  }
-});
-
-export default styles;
