+++ data/GiantsData.js
@@ -1,0 +1,128 @@
+export default [
+    {
+      id: 1,
+      name: "Teddy Friendly",
+      firstname: "Teddy",
+      location: "Høje Taastrup",
+      image: require("../assets/images/Teddy.png"),
+      desc:
+        "Teddy Friendly is a recycle sculpture Thomas Dambo made as a part of the larger project “Forgotten Giants”. The project consists of 6 large sculptures made in the outskirts of Copenhagen. Common for all sculptures is that they are made solely from local scrapwood and recycled materials, and made in a hidden beautiful location. This way it invites the viewers to go on a treasure hunt, not only to see the sculptures, but also to discover hidden gems in nature.\n\nTeddy friendly is made from local scrapwood, including cut offs from cut down local trees, which has been used for his fur. He is standing at a water stream, at a lake holding out his hand helping people to cross.\n\nThe Thomas Dambo team had help from a local activation center, which provided four unemployed people and a teacher to assist in building the sculpture. The teachers name was Teddy, he was a super friendly and hardworking guy, and together with his team he made a huge effort, therefore Thomas Dambo decided to name the sculpture after him. ",
+      region: {
+        latitude: 55.661788,
+        longitude: 12.266105,
+        latitudeDelta: 0.10022,
+        longitudeDelta: 0.101021
+      },
+      transport: {
+        station: "Høje Taastrup Station",
+        stog: "b, f",
+        bus: "120",
+        parking: "Hakkemosevej, 2630 Taastrup"
+      }
+    },
+    {
+      id: 2,
+      name: "Oscar Under The Bridge",
+      firstname: "Oscar",
+      location: "Ishøj",
+      image: require("../assets/images/Oscar.png"),
+      desc:
+        "Oscar Under The Bridge is a recycle sculpture Thomas Dambo made as a part of the larger project “Forgotten Giants”. The project consists of 6 large sculptures made in the outskirts of Copenhagen. Common for all sculptures is that they are made solely from local scrapwood and recycled materials, and made in a hidden beautiful location. This way it invites the viewers to go on a treasure hunt, not only to see the sculptures, but also to discover hidden gems in nature.\n\nOscar under the bridge is made from scrapwood from a torn down watermill and broken pallets from local industries. He is named after an artist from Chile who came to visit Thomas and help him in his work during the time that Thomas and his team was building the sculpture.",
+      region: {
+        latitude: 55.601398,
+        longitude: 12.372434,
+        latitudeDelta: 0.10022,
+        longitudeDelta: 0.10021
+      },
+      transport: {
+        station: "Ishøj station",
+        stog: "E, A",
+        bus: "128",
+        parking: "Skovvej 50, 2635 Ishøj or at Arken Museum"
+      }
+    },
+    {
+      id: 3,
+      name: "Sleeping Louis",
+      firstname: "Louis",
+      location: "Rødovre",
+      image: require("../assets/images/Louis.png"),
+      desc:
+        "Sleeping Louis is a recycle sculpture Thomas Dambo made as a part of the larger project “Forgotten Giants”. The project consists of 6 large sculptures made in the outskirts of Copenhagen. Common for all sculptures is that they are made solely from local scrapwood and recycled materials, and made in a hidden beautiful location. This way it invites the viewers to go on a treasure hunt, not only to see the sculptures, but also to discover hidden gems in nature.\n\nSleeping Louis is taking a nap on a hill, covered in trees and nature, in a secret place in Rødovre outside of Copenhagen. People can crawl into his gaping mouth, and play or even sleep inside of him.\n\nSleeping Louis is made solely from local scrapwood scavenged by the Thomas Dambo team. To build the sculpture the team had help from a group of young volunteers from an organisation helping young unemployed people get work experience before reentering the work market. Thomas’ old assistant Louis from Bristol, England came back to Copenhagen for a visit, and worked hard to help finish the sculpture, why Thomas and the team decided to name the sculpture after him.",
+      region: {
+        latitude: 55.674023,
+        longitude: 12.431133,
+        latitudeDelta: 0.10022,
+        longitudeDelta: 0.10021
+      },
+      transport: {
+        station: "Brøndbyøster Station",
+        stog: "B",
+        bus: "9A",
+        parking: "Korsdalsvej 112, 2610 Rødovre"
+      }
+    },
+    {
+      id: 4,
+      name: "Little Tilde",
+      firstname: "Tilde",
+      location: "Vallensbæk",
+      image: require("../assets/images/Tilde.png"),
+      desc:
+        "Little Tilde is a recycle sculpture Thomas Dambo made as a part of the larger project “Forgotten Giants”. The project consists of 6 large sculptures made in the outskirts of Copenhagen. Common for all sculptures is that they are made solely from local scrapwood and recycled materials, and made in a hidden beautiful location. This way it invites the viewers to go on a treasure hunt, not only to see the sculptures, but also to discover hidden gems in nature.\n\nLittle Tilde is located in the area of Vallensbæk Mose, af beautiful area filled with wild nature and animal life. She watches through the trees at the other side of a small lake, from where you will be able to catch a glimpse of her at a distance.\n\nTilde is made solely from local scrapwood scavenged by the Thomas Dambo team. To build the sculpture the team had help from a group of volunteers including two senior craftsmen (one of which was aged 86), and a couple of locals including Tilde, from whom the sculpture got her name.\n\nAs a little bonus Thomas Dambo made 28 birdhouses inside the sculpture, for the birds and maybe a squirrel to seek shelter when winter hits Vallensbæk Mose.",
+      region: {
+        latitude: 55.643548,
+        longitude: 12.345511,
+        latitudeDelta: 0.10022,
+        longitudeDelta: 0.10021
+      },
+      transport: {
+        station: "Vallensbæk Station",
+        stog: "A",
+        bus: "143, 144",
+        parking: "Rendsagervej, 2625 Vallensbæk"
+      }
+    },
+    {
+      id: 5,
+      name: "Thomas On The Mountain",
+      firstname: "Thomas",
+      location: "Albertslund",
+      image: require("../assets/images/Thomas.png"),
+      desc:
+        "Thomas On The Mountain is a recycle sculpture Thomas Dambo made as a part of the larger project “Forgotten Giants”. The project consists of 6 large sculptures made in the outskirts of Copenhagen. Common for all sculptures is that they are made solely from local scrapwood and recycled materials, and made in a hidden beautiful location. This way it invites the viewers to go on a treasure hunt, not only to see the sculptures, but also to discover hidden gems in nature.\n\nThomas On The Mountain has found a good place to chill on top of a hill. From here he has a great view of the municipality of Albertslund, and so can you if you join him. His legs are long, and a good place to sit with your friends.\n\nThe giant sculpture is made solely from local scrapwood scavenged by the Thomas Dambo team. To build the sculpture the team had help from a group of young volunteers from a local school as well as a couple of seniors. The giant is named after Thomas Dambos intern also named Thomas, who has been a huge part of the Forgotten Giants project, and helped create all six sculptures.",
+      region: {
+        latitude: 55.647515,
+        longitude: 12.342947,
+        latitudeDelta: 0.10022,
+        longitudeDelta: 0.10021
+      },
+      transport: {
+        station: "Albertslund Station",
+        stog: "B",
+        bus: "117",
+        parking: "Vridsløsevej, 2620 Vallensbæk"
+      }
+    },
+    {
+      id: 6,
+      name: "Hill Top Trine",
+      firstname: "Trine",
+      location: "Avedøre",
+      image: require("../assets/images/Trine.png"),
+      desc:
+        "Hill Top Trine is a recycle sculpture Thomas Dambo made as a part of the larger project “Forgotten Giants”. The project consists of 6 large sculptures made in the outskirts of Copenhagen. Common for all sculptures is that they are made solely from local scrapwood and recycled materials, and made in a hidden beautiful location. This way it invites the viewers to go on a treasure hunt, not only to see the sculptures, but also to discover hidden gems in nature.\n\nHill Top Trine is resting on top of a little hill in Hvidovre, Copenhagen, where you can crawl into the palms of her hands and get a beautiful viewpoint overlooking Avedøresletten. She is named after one of the volunteers who worked on one of the other hidden giants.\n\nJust like her siblings, Hill Top Trine is made from scrapwood, branches and old pallets from Stark. The Thomas Dambo team have had help building the wooden sculpture from a handful of local volunteers.",
+      region: {
+        latitude: 55.633884,
+        longitude: 12.441283,
+        latitudeDelta: 0.10022,
+        longitudeDelta: 0.10021
+      },
+      transport: {
+        station: "Avedøre Station",
+        stog: "A",
+        bus: "1A, 4A",
+        parking: "Filmbyen 34, 2650 Hvidovre"
+      }
+    }
+  ];
