+++ screens/RewardScreen/RewardScreen.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react";
 import { Text, View, ImageBackground, ScrollView } from "react-native";
-import Styles from "./RewardScreenStyles";
 import { YellowButton } from "../../components/Buttons/YellowButton";
 import { DefaultButton } from "../../components/Buttons/DefaultButton";
 import { BigReward } from "../../components/Reward/BigReward";
-import rewards from "../RewardCollection/RewardData";
+import RewardData from "../../data/RewardData";
+import Styles from "./RewardStyles";
 
 export default class RewardScreen extends Component {
   componentWillMount() {
-    rewards.map(reward => {
+    RewardData.map(reward => {
       if (this.props.navigation.getParam("giantId") === reward.id)
         reward.found = true;
       reward.date = this.props.navigation.getParam("date");
