+++ screens/RewardScreen/RewardStyles.js
